refactor(webapp): split admin back-office org page into per-section components
