+++ apps/webapp/app/routes/admin.back-office.orgs.$orgId.tsx
@@ -1,102 +1,35 @@
-import { Form, useNavigation, useSearchParams } from "@remix-run/react";
+import { useNavigation, useSearchParams } from "@remix-run/react";
 import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/server-runtime";
-import { useEffect, useState } from "react";
-import { redirect, typedjson, useTypedActionData, useTypedLoaderData } from "remix-typedjson";
-import { z } from "zod";
-import { Button, LinkButton } from "~/components/primitives/Buttons";
-import { CopyableText } from "~/components/primitives/CopyableText";
-import { FormError } from "~/components/primitives/FormError";
-import { Header1, Header2 } from "~/components/primitives/Headers";
-import { Input } from "~/components/primitives/Input";
-import { Label } from "~/components/primitives/Label";
-import { Paragraph } from "~/components/primitives/Paragraph";
-import * as Property from "~/components/primitives/PropertyTable";
-import { prisma } from "~/db.server";
-import { env } from "~/env.server";
+import { useEffect } from "react";
 import {
-  RateLimitTokenBucketConfig,
-  RateLimiterConfig,
-} from "~/services/authorizationRateLimitMiddleware.server";
-import { logger } from "~/services/logger.server";
-import { type Duration } from "~/services/rateLimiter.server";
+  redirect,
+  typedjson,
+  useTypedActionData,
+  useTypedLoaderData,
+} from "remix-typedjson";
+import {
+  MAX_PROJECTS_INTENT,
+  MAX_PROJECTS_SAVED_VALUE,
+  MaxProjectsSection,
+} from "~/components/admin/backOffice/MaxProjectsSection";
+import { handleMaxProjectsAction } from "~/components/admin/backOffice/MaxProjectsSection.server";
+import {
+  RATE_LIMIT_INTENT,
+  RATE_LIMIT_SAVED_VALUE,
+  RateLimitSection,
+} from "~/components/admin/backOffice/RateLimitSection";
+import {
+  handleRateLimitAction,
+  resolveEffectiveRateLimit,
+} from "~/components/admin/backOffice/RateLimitSection.server";
+import { LinkButton } from "~/components/primitives/Buttons";
+import { CopyableText } from "~/components/primitives/CopyableText";
+import { Header1 } from "~/components/primitives/Headers";
+import { Paragraph } from "~/components/primitives/Paragraph";
+import { prisma } from "~/db.server";
 import { requireUser } from "~/services/session.server";
 
 const SAVED_QUERY_KEY = "saved";
-const SAVED_QUERY_VALUE = "1";
-
-type EffectiveRateLimit = {
-  source: "override" | "default";
-  config: RateLimiterConfig;
-};
-
-function systemDefaultRateLimit(): RateLimiterConfig {
-  return {
-    type: "tokenBucket",
-    refillRate: env.API_RATE_LIMIT_REFILL_RATE,
-    interval: env.API_RATE_LIMIT_REFILL_INTERVAL as Duration,
-    maxTokens: env.API_RATE_LIMIT_MAX,
-  };
-}
-
-function resolveEffectiveRateLimit(override: unknown): EffectiveRateLimit {
-  if (override == null) {
-    return { source: "default", config: systemDefaultRateLimit() };
-  }
-  const parsed = RateLimiterConfig.safeParse(override);
-  if (parsed.success) {
-    return { source: "override", config: parsed.data };
-  }
-  // Column holds malformed JSON — fall back silently. Admin must investigate
-  // at the DB level; this UI can't recover it.
-  return { source: "default", config: systemDefaultRateLimit() };
-}
-
-function parseDurationToMs(duration: string): number {
-  const match = duration.trim().match(/^(\d+)\s*(ms|s|m|h|d)$/);
-  if (!match) return 0;
-  const value = parseInt(match[1], 10);
-  switch (match[2]) {
-    case "ms":
-      return value;
-    case "s":
-      return value * 1_000;
-    case "m":
-      return value * 60_000;
-    case "h":
-      return value * 3_600_000;
-    case "d":
-      return value * 86_400_000;
-    default:
-      return 0;
-  }
-}
-
-function describeRateLimit(
-  refillRate: number,
-  intervalMs: number,
-  maxTokens: number
-): { sustained: string; burst: string } | null {
-  if (refillRate <= 0 || intervalMs <= 0 || maxTokens <= 0) return null;
-  const perMin = (refillRate * 60_000) / intervalMs;
-  let sustained: string;
-  if (perMin >= 1) {
-    sustained = `${Math.round(perMin).toLocaleString()} requests per minute`;
-  } else {
-    const perHour = perMin * 60;
-    if (perHour >= 1) {
-      sustained = `${Math.round(perHour).toLocaleString()} requests per hour`;
-    } else {
-      const perDay = perHour * 24;
-      const formatted =
-        perDay >= 10 ? Math.round(perDay).toLocaleString() : perDay.toFixed(1);
-      sustained = `${formatted} requests per day`;
-    }
-  }
-  return {
-    sustained,
-    burst: `${maxTokens.toLocaleString()} request burst allowance`,
-  };
-}
 
 export async function loader({ request, params }: LoaderFunctionArgs) {
   const user = await requireUser(request);
@@ -117,6 +50,7 @@ export async function loader({ request, params }: LoaderFunctionArgs) {
       title: true,
       createdAt: true,
       apiRateLimiterConfig: true,
+      maximumProjectCount: true,
     },
   });
 
@@ -126,23 +60,8 @@ export async function loader({ request, params }: LoaderFunctionArgs) {
 
   const effective = resolveEffectiveRateLimit(org.apiRateLimiterConfig);
 
-  return typedjson({
-    org,
-    effective,
-  });
+  return typedjson({ org, effective });
 }
-
-const SetRateLimitSchema = z.object({
-  intent: z.literal("set-rate-limit"),
-  refillRate: z.coerce.number().int().min(1),
-  interval: z
-    .string()
-    .trim()
-    .refine((v) => parseDurationToMs(v) > 0, {
-      message: "Must be a duration like 10s, 1m, 500ms.",
-    }),
-  maxTokens: z.coerce.number().int().min(1),
-});
 
 export async function action({ request, params }: ActionFunctionArgs) {
   const user = await requireUser(request);
@@ -156,50 +75,37 @@ export async function action({ request, params }: ActionFunctionArgs) {
   }
 
   const formData = await request.formData();
-  const submission = SetRateLimitSchema.safeParse(Object.fromEntries(formData));
-  if (!submission.success) {
-    return typedjson(
-      { errors: submission.error.flatten().fieldErrors },
-      { status: 400 }
+  const intent = formData.get("intent");
+
+  if (intent === MAX_PROJECTS_INTENT) {
+    const result = await handleMaxProjectsAction(formData, orgId, user.id);
+    if (!result.ok) {
+      return typedjson(
+        { section: MAX_PROJECTS_SAVED_VALUE, errors: result.errors },
+        { status: 400 }
+      );
+    }
+    return redirect(
+      `/admin/back-office/orgs/${orgId}?${SAVED_QUERY_KEY}=${MAX_PROJECTS_SAVED_VALUE}`
     );
   }
 
-  const existing = await prisma.organization.findFirst({
-    where: { id: orgId },
-    select: { apiRateLimiterConfig: true },
-  });
-  if (!existing) {
-    throw new Response(null, { status: 404 });
-  }
-
-  const built = RateLimitTokenBucketConfig.safeParse({
-    type: "tokenBucket",
-    refillRate: submission.data.refillRate,
-    interval: submission.data.interval,
-    maxTokens: submission.data.maxTokens,
-  });
-  if (!built.success) {
-    return typedjson(
-      { errors: built.error.flatten().fieldErrors },
-      { status: 400 }
+  if (intent === RATE_LIMIT_INTENT) {
+    const result = await handleRateLimitAction(formData, orgId, user.id);
+    if (!result.ok) {
+      return typedjson(
+        { section: RATE_LIMIT_SAVED_VALUE, errors: result.errors },
+        { status: 400 }
+      );
+    }
+    return redirect(
+      `/admin/back-office/orgs/${orgId}?${SAVED_QUERY_KEY}=${RATE_LIMIT_SAVED_VALUE}`
     );
   }
-  const next = built.data;
 
-  await prisma.organization.update({
-    where: { id: orgId },
-    data: { apiRateLimiterConfig: next as any },
-  });
-
-  logger.info("admin.backOffice.rateLimit", {
-    adminUserId: user.id,
-    orgId,
-    previous: existing.apiRateLimiterConfig,
-    next,
-  });
-
-  return redirect(
-    `/admin/back-office/orgs/${orgId}?${SAVED_QUERY_KEY}=${SAVED_QUERY_VALUE}`
+  return typedjson(
+    { section: null, errors: { intent: ["Unknown intent."] } },
+    { status: 400 }
   );
 }
 
@@ -209,47 +115,19 @@ export default function BackOfficeOrgPage() {
   const navigation = useNavigation();
   const isSubmitting = navigation.state !== "idle";
 
+  const errorSection =
+    actionData && "section" in actionData ? actionData.section : null;
   const errors =
-    actionData && "errors" in actionData ? actionData.errors : null;
-  const hasFieldErrors =
-    !!errors && typeof errors === "object" && Object.keys(errors).length > 0;
-  const fieldError = (field: string) =>
-    errors && typeof errors === "object" && field in errors
-      ? (errors as Record<string, string[] | undefined>)[field]?.[0]
-      : undefined;
-
-  const current =
-    effective.config.type === "tokenBucket" ? effective.config : null;
-
-  const [isEditing, setIsEditing] = useState(false);
-  const [refillRate, setRefillRate] = useState(
-    current ? String(current.refillRate) : ""
-  );
-  const [intervalStr, setIntervalStr] = useState(
-    current ? String(current.interval) : ""
-  );
-  const [maxTokens, setMaxTokens] = useState(
-    current ? String(current.maxTokens) : ""
-  );
+    actionData && "errors" in actionData
+      ? (actionData.errors as Record<string, string[] | undefined>)
+      : null;
 
   const [searchParams, setSearchParams] = useSearchParams();
-  const savedJustNow = searchParams.get(SAVED_QUERY_KEY) === SAVED_QUERY_VALUE;
-
-  // If a submit comes back with validation errors, re-open edit mode so the
-  // admin can see and correct them without clicking Edit again.
-  useEffect(() => {
-    if (hasFieldErrors) setIsEditing(true);
-  }, [hasFieldErrors]);
-
-  // On successful save, drop back to view mode (the component stays mounted
-  // across the same-route redirect, so `isEditing` wouldn't reset on its own).
-  useEffect(() => {
-    if (savedJustNow) setIsEditing(false);
-  }, [savedJustNow]);
+  const savedSection = searchParams.get(SAVED_QUERY_KEY);
 
   // Auto-dismiss the "saved" banner after a few seconds.
   useEffect(() => {
-    if (!savedJustNow) return;
+    if (!savedSection) return;
     const t = setTimeout(() => {
       setSearchParams(
         (prev) => {
@@ -260,28 +138,7 @@ export default function BackOfficeOrgPage() {
       );
     }, 3000);
     return () => clearTimeout(t);
-  }, [savedJustNow, setSearchParams]);
-
-  const currentDescription = current
-    ? describeRateLimit(
-        current.refillRate,
-        parseDurationToMs(String(current.interval)),
-        current.maxTokens
-      )
-    : null;
-
-  const previewDescription = describeRateLimit(
-    Number(refillRate) || 0,
-    parseDurationToMs(intervalStr),
-    Number(maxTokens) || 0
-  );
-
-  const cancelEdit = () => {
-    setRefillRate(current ? String(current.refillRate) : "");
-    setIntervalStr(current ? String(current.interval) : "");
-    setMaxTokens(current ? String(current.maxTokens) : "");
-    setIsEditing(false);
-  };
+  }, [savedSection, setSearchParams]);
 
   return (
     <div className="flex flex-col gap-6 py-4">
@@ -297,156 +154,19 @@ export default function BackOfficeOrgPage() {
         </LinkButton>
       </div>
 
-      <section className="flex flex-col gap-3 rounded-md border border-charcoal-700 bg-charcoal-800 p-4">
-        <div className="flex items-center justify-between">
-          <Header2>API rate limit</Header2>
-          {!isEditing && (
-            <Button
-              variant="tertiary/small"
-              onClick={() => setIsEditing(true)}
-              disabled={isSubmitting || effective.config.type !== "tokenBucket"}
-            >
-              Edit
-            </Button>
-          )}
-        </div>
+      <RateLimitSection
+        effective={effective}
+        errors={errorSection === RATE_LIMIT_SAVED_VALUE ? errors : null}
+        savedJustNow={savedSection === RATE_LIMIT_SAVED_VALUE}
+        isSubmitting={isSubmitting}
+      />
 
-        {savedJustNow && (
-          <div className="rounded-md border border-green-600/40 bg-green-600/10 px-3 py-2">
-            <Paragraph variant="small" className="text-green-500">
-              Rate limit saved.
-            </Paragraph>
-          </div>
-        )}
-
-        <Paragraph variant="small">
-          Status:{" "}
-          {effective.source === "override"
-            ? "Custom override active."
-            : "Using system default."}
-        </Paragraph>
-
-        {!isEditing ? (
-          <>
-            <Property.Table>
-              {effective.config.type === "tokenBucket" ? (
-                currentDescription ? (
-                  <>
-                    <Property.Item>
-                      <Property.Label>Sustained rate</Property.Label>
-                      <Property.Value>{currentDescription.sustained}</Property.Value>
-                    </Property.Item>
-                    <Property.Item>
-                      <Property.Label>Burst allowance</Property.Label>
-                      <Property.Value>{currentDescription.burst}</Property.Value>
-                    </Property.Item>
-                  </>
-                ) : (
-                  <Property.Item>
-                    <Property.Value>
-                      Invalid interval on the stored config.
-                    </Property.Value>
-                  </Property.Item>
-                )
-              ) : (
-                <>
-                  <Property.Item>
-                    <Property.Label>Type</Property.Label>
-                    <Property.Value>{effective.config.type}</Property.Value>
-                  </Property.Item>
-                  <Property.Item>
-                    <Property.Label>Window</Property.Label>
-                    <Property.Value>{String(effective.config.window)}</Property.Value>
-                  </Property.Item>
-                  <Property.Item>
-                    <Property.Label>Tokens</Property.Label>
-                    <Property.Value>
-                      {effective.config.tokens.toLocaleString()}
-                    </Property.Value>
-                  </Property.Item>
-                </>
-              )}
-            </Property.Table>
-            {effective.config.type !== "tokenBucket" && (
-              <Paragraph variant="small" className="text-amber-500">
-                This override is a {effective.config.type} limit and can't be
-                edited from this form. Change it in the database directly.
-              </Paragraph>
-            )}
-          </>
-        ) : (
-          <Form method="post" className="flex flex-col gap-3 pt-2">
-            <input type="hidden" name="intent" value="set-rate-limit" />
-
-            <div className="flex flex-col gap-1">
-              <Label>Refill rate (tokens per interval)</Label>
-              <Input
-                name="refillRate"
-                type="number"
-                min={1}
-                value={refillRate}
-                onChange={(e) => setRefillRate(e.target.value)}
-                required
-              />
-              <FormError>{fieldError("refillRate")}</FormError>
-            </div>
-
-            <div className="flex flex-col gap-1">
-              <Label>Interval (e.g. 10s, 1m)</Label>
-              <Input
-                name="interval"
-                type="text"
-                value={intervalStr}
-                onChange={(e) => setIntervalStr(e.target.value)}
-                required
-              />
-              <FormError>{fieldError("interval")}</FormError>
-            </div>
-
-            <div className="flex flex-col gap-1">
-              <Label>Max tokens (burst allowance)</Label>
-              <Input
-                name="maxTokens"
-                type="number"
-                min={1}
-                value={maxTokens}
-                onChange={(e) => setMaxTokens(e.target.value)}
-                required
-              />
-              <FormError>{fieldError("maxTokens")}</FormError>
-            </div>
-
-            <Paragraph variant="small" className="text-text-dimmed">
-              {previewDescription
-                ? `Preview: ${previewDescription.sustained} · ${previewDescription.burst}.`
-                : "Preview: enter valid values to see the effective limit."}
-            </Paragraph>
-
-            <div className="flex items-center gap-2">
-              <Button
-                type="submit"
-                variant="primary/medium"
-                disabled={
-                  isSubmitting ||
-                  !refillRate.trim() ||
-                  !intervalStr.trim() ||
-                  !maxTokens.trim()
-                }
-              >
-                Save
-              </Button>
-              <Button
-                type="button"
-                variant="tertiary/medium"
-                onClick={cancelEdit}
-                disabled={isSubmitting}
-              >
-                Cancel
-              </Button>
-            </div>
-          </Form>
-        )}
-      </section>
+      <MaxProjectsSection
+        maximumProjectCount={org.maximumProjectCount}
+        errors={errorSection === MAX_PROJECTS_SAVED_VALUE ? errors : null}
+        savedJustNow={savedSection === MAX_PROJECTS_SAVED_VALUE}
+        isSubmitting={isSubmitting}
+      />
     </div>
   );
 }
